+++ spec/app/domain/issuers/issuer_types/issuer_base_type_spec.rb
@@ -1,23 +1,20 @@
 # frozen_string_literal: true
 require 'spec_helper'
 
-class BaseTypeTest < PlatformBaseType
+class BaseTypeTest < IssuerBaseType
   def validate(params)
-    super
-  end
-
-  def default_secret_method
     super
   end
 end
 
-describe "PlatformBaseType input validation" do
+describe "IssuerBaseType input validation" do
   
   context "when all base input is valid" do
     it "then the input validation succeeds" do
-      params = ActionController::Parameters.new(id: "aws-platform-1", 
+      params = ActionController::Parameters.new(id: "aws-issuer-1",
                                                 max_ttl: 2000, 
-                                                type: "aws")
+                                                type: "aws",
+                                                data: {})
       expect { BaseTypeTest.new.validate(params) }
         .to_not raise_error
     end
@@ -25,9 +22,10 @@ describe "PlatformBaseType input validation" do
 
   context "when max_ttl is a negative number" do
     it "then the input validation fails" do
-      params = ActionController::Parameters.new(id: "aws-platform-1", 
+      params = ActionController::Parameters.new(id: "aws-issuer-1",
                                                 max_ttl: -2000, 
-                                                type: "aws")
+                                                type: "aws",
+                                                data: {})
       expect { BaseTypeTest.new.validate(params) }
         .to raise_error(ApplicationController::BadRequest)
     end
@@ -35,9 +33,10 @@ describe "PlatformBaseType input validation" do
 
   context "when max_ttl is 0" do
     it "then the input validation fails" do
-      params = ActionController::Parameters.new(id: "aws-platform-1", 
+      params = ActionController::Parameters.new(id: "aws-issuer-1",
                                                 max_ttl: 0, 
-                                                type: "aws")
+                                                type: "aws",
+                                                data: {})
       expect { BaseTypeTest.new.validate(params) }
         .to raise_error(ApplicationController::BadRequest)
     end
@@ -45,9 +44,10 @@ describe "PlatformBaseType input validation" do
 
   context "when max_ttl is a floating number" do
     it "then the input validation fails" do
-      params = ActionController::Parameters.new(id: "aws-platform-1", 
+      params = ActionController::Parameters.new(id: "aws-issuer-1",
                                                 max_ttl: 4.5, 
-                                                type: "aws")
+                                                type: "aws",
+                                                data: {})
       expect { BaseTypeTest.new.validate(params) }
         .to raise_error(ApplicationController::BadRequest)
     end
@@ -57,7 +57,8 @@ describe "PlatformBaseType input validation" do
     it "then the input validation fails" do
       params = ActionController::Parameters.new(id: 1, 
                                                 max_ttl: 2000, 
-                                                type: "aws")
+                                                type: "aws",
+                                                data: {})
       expect { BaseTypeTest.new.validate(params) }
         .to raise_error(ApplicationController::BadRequest)
     end
@@ -67,7 +68,8 @@ describe "PlatformBaseType input validation" do
     it "then the input validation fails" do
       params = ActionController::Parameters.new(id: "a" * 61,
                                                 max_ttl: 2000,
-                                                type: "aws")
+                                                type: "aws",
+                                                data: {})
       expect { BaseTypeTest.new.validate(params) }
         .to raise_error(ApplicationController::BadRequest)
     end
@@ -77,7 +79,8 @@ describe "PlatformBaseType input validation" do
     it "then the input validation fails" do
       params = ActionController::Parameters.new(id: "a^", 
                                                 max_ttl: 2000, 
-                                                type: "aws")
+                                                type: "aws",
+                                                data: {})
       expect { BaseTypeTest.new.validate(params) }
         .to raise_error(ApplicationController::BadRequest)
     end
@@ -87,8 +90,21 @@ describe "PlatformBaseType input validation" do
     it "then the input validation fails" do
       params = ActionController::Parameters.new(id: "a hf", 
                                                 max_ttl: 2000, 
-                                                type: "aws")
+                                                type: "aws",
+                                                data: {})
       expect { BaseTypeTest.new.validate(params) }
+        .to raise_error(ApplicationController::BadRequest)
+    end
+  end
+
+  context "when invalid parameter is added to the body main section" do
+    it "then the input validation fails" do
+      params = ActionController::Parameters.new(id: "aws-issuer-1",
+                                                max_ttl: 2000,
+                                                type: "aws",
+                                                invalid_param: "a",
+                                                data: {})
+      expect { AwsIssuerType.new.validate(params) }
         .to raise_error(ApplicationController::BadRequest)
     end
   end