+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   (and update the corresponding date), or add a new version.
 
 ## [1.0.6-cloud] - 2023-08-21
+### Changed
+- Renamed `platforms` to `issuers`, changed internal structure of ephemeral secret requests and removed default issuer secret
+  [ONYX-42993](https://ca-il-jira.il.cyber-ark.com:8443/browse/ONYX-42993)
 
 ## [1.0.5-cloud] - 2023-08-16
 ### Security